+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 - Adds support for groups on dynamic views (Show group also in dynamic view #1).
 - Fixes an issue when nested groups are configured, but groups are not nested.
+- StructurizrPlantUMLExporter: Adds support for group icons.
 
 ## 1.12.1 (12th March 2023)